# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SubTrack – Subscription Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List all subscriptions for the current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived, all]
          description: Filter by status
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
          description: Filter by category ID
        - name: billingCycle
          in: query
          schema:
            type: string
            enum: [weekly, monthly, quarterly, semi_annual, yearly]
          description: Filter by billing cycle
        - name: search
          in: query
          schema:
            type: string
          description: Search by name, notes
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
          description: Minimum price filter
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
          description: Maximum price filter
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, price, renewalDate, createdAt, monthlyEquivalent]
          description: Sort field
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
          description: Sort direction
      responses:
        "200":
          description: List of subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
        "401":
          description: Unauthorized

    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Create a new subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "201":
          description: Created subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /subscriptions/{id}:
    get:
      operationId: getSubscription
      tags: [subscriptions]
      summary: Get a subscription by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "404":
          description: Not found

    patch:
      operationId: updateSubscription
      tags: [subscriptions]
      summary: Update a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionUpdate"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "404":
          description: Not found

    delete:
      operationId: deleteSubscription
      tags: [subscriptions]
      summary: Delete a subscription permanently
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /subscriptions/{id}/archive:
    patch:
      operationId: archiveSubscription
      tags: [subscriptions]
      summary: Archive a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Archived subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}/restore:
    patch:
      operationId: restoreSubscription
      tags: [subscriptions]
      summary: Restore an archived subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

  # ──────────────────────────────
  # CATEGORIES
  # ──────────────────────────────
...
components:
  schemas:
...
    Subscription:
      type: object
      required: [id, userId, name, price, currency, billingCycle, renewalDate, isActive, isArchived, monthlyEquivalent, annualEquivalent, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        price:
          type: number
        currency:
          type: string
        billingCycle:
          type: string
          enum: [weekly, monthly, quarterly, semi_annual, yearly]
        renewalDate:
          type: string
          description: ISO date string YYYY-MM-DD
        paymentMethod:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
        isArchived:
          type: boolean
        monthlyEquivalent:
          type: number
          description: Monthly cost equivalent in the original currency
        annualEquivalent:
          type: number
          description: Annual cost equivalent in the original currency
        daysUntilRenewal:
          type: ["integer", "null"]
          description: Days until next renewal (null if archived)
        createdAt:
          type: string
        updatedAt:
          type: string
    SubscriptionInput:
      type: object
      required: [name, price, currency, billingCycle, renewalDate]
      properties:
        name:
          type: string
          minLength: 1
        logoUrl:
          type: string
        categoryId:
          type: integer
        price:
          type: number
          minimum: 0
        currency:
          type: string
          default: USD
        billingCycle:
          type: string
          enum: [weekly, monthly, quarterly, semi_annual, yearly]
        renewalDate:
          type: string
          description: ISO date YYYY-MM-DD
        paymentMethod:
          type: string
        notes:
          type: string
        isActive:
          type: boolean
          default: true
    SubscriptionUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        logoUrl:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        price:
          type: number
          minimum: 0
        currency:
          type: string
        billingCycle:
          type: string
          enum: [weekly, monthly, quarterly, semi_annual, yearly]
        renewalDate:
          type: string
        paymentMethod:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean

    # ── Category ──
...